Fixed some bugs.

+++ include/benchmarks/gemm/gemm_custom.h
@@ -12,8 +12,8 @@ static void gemm_custom(benchmark::State& state) {
 
     for (int i = 0; i < SIZE; i++) {
         for (int j = 0; j < SIZE; j++) {
-            a[i * SIZE + j] = randomDouble(-1.0, 1.0);
-            b[i * SIZE + j] = randomDouble(-1.0, 1.0);
+            a[i * SIZE + j] = randomFloat(-1.0, 1.0);
+            b[i * SIZE + j] = randomFloat(-1.0, 1.0);
         }
     }
