+++ include/benchmarks/similarity/similarity_blaze.h
@@ -7,12 +7,12 @@
 #include "benchmarks/shared/random.hpp"
 
 static void similarity_blaze(benchmark::State& state) {
-    blaze::StaticMatrix<double, SIZE, SIZE, blaze::rowMajor> m1, m2, res;
+    blaze::StaticMatrix<float, SIZE, SIZE, blaze::rowMajor> m1, m2, res;
 
     m1 =
-        blaze::generate(SIZE, SIZE, [](std::size_t i, std::size_t j) { return randomDouble(-1.0, 1.0); });
+        blaze::generate(SIZE, SIZE, [](std::size_t i, std::size_t j) { return randomFloat(-1.0, 1.0); });
     m2 =
-        blaze::generate(SIZE, SIZE, [](std::size_t i, std::size_t j) { return randomDouble(-1.0, 1.0); });
+        blaze::generate(SIZE, SIZE, [](std::size_t i, std::size_t j) { return randomFloat(-1.0, 1.0); });
 
     for (auto _ : state) {
         benchmark::DoNotOptimize(res);
