+++ include/benchmarks/similarity/similarity_eigen.h
@@ -6,10 +6,10 @@
 #include "benchmarks/shared/common.hpp"
 
 static void similarity_eigen(benchmark::State& state) {
-    Eigen::Matrix<double, SIZE, SIZE> m1, m2, res;
+    Eigen::Matrix<float, SIZE, SIZE> m1, m2, res;
 
-    m1 = Eigen::Matrix<double, SIZE, SIZE>::Random();
-    m2 = Eigen::Matrix<double, SIZE, SIZE>::Random();
+    m1 = Eigen::Matrix<float, SIZE, SIZE>::Random();
+    m2 = Eigen::Matrix<float, SIZE, SIZE>::Random();
 
     for (auto _ : state) {
         benchmark::DoNotOptimize(res);
